pyproject.toml
[project]
name = "reclaimerr"
version = "0.1.0-alpha1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "apprise>=1.9.7",
    "apscheduler>=3.11.2,<4.0.0",
    "argon2-cffi>=23.1.0",
    "cryptography>=46.0.5",
    "fastapi[standard]>=0.135.1",
    "granian>=2.7.2",
    "niquests[speedups]>=3.16.1",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "semver>=3.0.4",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.45",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = ["ruff", "uvicorn>=0.41.0"]
desktop = [
    "pyinstaller",
    "pystray",
    "platformdirs>=4.9.6",
    "filelock>=3.16.0",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/core", "backend/api", "backend/services"]

[tool.ruff]
line-length = 88
target-version = "py39"
include = [
    "backend/core/**/*.py",
    "backend/api/**/*.py",
    "backend/services/**/*.py",
]

[tool.ruff.lint]
select = ["I", "UP"]

[tool.basedpyright]
include = ["backend"]
typeCheckingMode = "standard"
